Screw that complex logic, here's solid base logic
